Resolved against post-#190 + post-#194 main:

1. resolve_at visibility — main has `pub(super)`, took main's.
2. ABI ID collision — main's just-landed sys_unlinkat=0x1_00B0 +
   sys_mkdirat=0x1_00B1 (#194's in-PR fix-up) collide with this PR's
   sys_fstatat=0x1_00B1 + sys_readlinkat=0x1_00B2. Bumped to:
     sys_fstatat:   0x1_00B1 → 0x1_00B2
     sys_readlinkat: 0x1_00B2 → 0x1_00B3
   Append-only convention preserved (still strictly above
   sys_mkdirat=0x1_00B1).
3. fs.rs / mod.rs import + dispatch arm conflicts auto-merge OK
   except for the dueling-add hunks — kept the superset (S1's
   unlinkat/mkdirat from main + S2's fstatat/readlinkat from this
   branch).
4. tests.rs EOF — took main's tests verbatim (now has #190 S0 +
   #194 S1 blocks), re-appended only this PR's S2 block.

511/511 lib tests pass; fmt + clippy clean.

+++ packages/kernel-wasm/src/vfs.rs
@@ -428,7 +428,7 @@ impl MountTable {
     pub fn set_len(&mut self, mount_id: MountId, inode: u64, length: u64) -> i32 {
         match self.mounts.get_mut(mount_id as usize) {
             Some(m) => m.backend.set_len(inode, length),
-            None => -(crate::abi::EBADF as i32),
+            None => -crate::abi::EBADF,
         }
     }
 
@@ -823,13 +823,13 @@ impl VfsBackend for RamfsBackend {
 
     fn set_len(&mut self, inode: u64, length: u64) -> i32 {
         let Some(content) = self.inodes.get_mut(&inode) else {
-            return -(crate::abi::EBADF as i32);
+            return -crate::abi::EBADF;
         };
         // usize::MAX on wasm32 is 4 GiB - 1; refuse anything that
         // would overflow the host buffer. EFBIG matches POSIX intent
         // (resource limit / file too large) better than EINVAL here.
         let Ok(len_usize) = usize::try_from(length) else {
-            return -(crate::abi::EFBIG as i32);
+            return -crate::abi::EFBIG;
         };
         content.resize(len_usize, 0);
         0
@@ -2811,7 +2811,7 @@ impl VfsBackend for OverlayBackend {
             // `truncate` for now. A future overlay rewrite that copies
             // up on write should also copy up on set_len. Issue #87.
             Some(&(Layer::Lower, _)) => -30, // -EROFS
-            None => -(crate::abi::EBADF as i32),
+            None => -crate::abi::EBADF,
         }
     }
